+++ harness/determined/core/__init__.py
@@ -1,5 +1,9 @@
 from determined.core._tensorboard_mode import TensorboardMode
-from determined.core._distributed import DistributedContext, DummyDistributedContext
+from determined.core._distributed import (
+    DistributedContext,
+    DummyDistributedContext,
+    _run_on_rank_0_and_broadcast,
+)
 from determined.core._checkpoint import (
     CheckpointContext,
     DownloadMode,
